Delete publications only after successful tests

+++ tests/static_publications/data.py
@@ -39,17 +39,19 @@ def assert_publication_before_post(workspace, publ_type, publication):
 
 
 @pytest.fixture(scope="session", autouse=True)
-def ensure_test_data(liferay_mock):
+def ensure_test_data(liferay_mock, request):
     # pylint: disable=unused-argument
     yield
-    process.ensure_layman_function(process.LAYMAN_DEFAULT_SETTINGS)
-    with app.app_context():
-        info = util.get_publication_infos()
 
-    for workspace, publ_type, publication in info:
-        headers = data.HEADERS.get(data.PUBLICATIONS[(workspace, publ_type, publication)][data.TEST_DATA].get('users_can_write', [None])[0])
-        process_client.delete_workspace_publication(publ_type, workspace, publication, headers=headers)
-        assert_publication_after_delete(workspace, publ_type, publication)
+    if request.node.testsfailed == 0:
+        process.ensure_layman_function(process.LAYMAN_DEFAULT_SETTINGS)
+        with app.app_context():
+            info = util.get_publication_infos()
+
+        for workspace, publ_type, publication in info:
+            headers = data.HEADERS.get(data.PUBLICATIONS[(workspace, publ_type, publication)][data.TEST_DATA].get('users_can_write', [None])[0])
+            process_client.delete_workspace_publication(publ_type, workspace, publication, headers=headers)
+            assert_publication_after_delete(workspace, publ_type, publication)
 
 
 def ensure_publication(workspace, publ_type, publication):
